chore: Rename Left Sidebar (#1113)
