+++ client/src/components/LeftSidebar/LeftSidebar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Categorical from "../categorical";
-import Continuous from "../continuous/continuous";
+import Categorical from "./components/Categorical/Categorical";
+import Continuous from "./components/Continuous/Continuous";
 import { LeftSidebarContainer, LeftSidebarWrapper } from "./style";
 
 function LeftSideBar() {
