+++ client/src/components/LeftSidebar/components/Categorical/categorical.css
