+++ client/src/components/LeftSidebar/components/Categorical/components/Category/components/CategoryValue/components/MiniHistogram/MiniHistogram.tsx
