+++ client/src/components/LeftSidebar/components/Categorical/components/Category/components/CategoryValue/labelUtil.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Colors } from "@blueprintjs/core";
 
-import { AnnotationsHelpers } from "../../util/stateManager";
+import { AnnotationsHelpers } from "util/stateManager";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any -- - FIXME: disabled temporarily on migrate to TS.
 export function isLabelErroneous(label: any, metadataField: any, schema: any) {
