+++ client/src/components/LeftSidebar/components/Categorical/constants.ts
