+++ client/src/components/LeftSidebar/components/Continuous/parallelCoordinates.css
