+++ client/src/components/LeftSidebar/components/Continuous/setupParallelCoordinates.ts
