+++ client/src/components/NavBar/infoMenu.tsx
