+++ client/src/components/framework/layoutSkeleton.tsx
@@ -5,7 +5,7 @@ import React from "react";
 /* App dependencies */
 import Controls from "common/components/Controls/Controls";
 import Layout from "./layout";
-import LeftSidebarSkeleton from "../leftSidebar/leftSidebarSkeleton";
+import LeftSidebarSkeleton from "../LeftSidebar/LeftSidebarSkeleton";
 import RightSidebarSkeleton from "../rightSidebar/rightSidebarSkeleton";
 
 /**
